requirements.txt
adafruit-circuitpython-ssd1306
adafruit-gpio
python-mpd2
Pillow
